Clean up the scripts and add documentation

+++ testing/configure/linux_install_lme.sh
@@ -16,8 +16,13 @@ done
 
 # Download a copy of the LME files
 sudo git clone https://github.com/cisagov/lme.git /opt/lme/
+
 # Execute script with root privileges
 sudo ./linux_install_lme.exp
 
-sudo cp /opt/lme/files_for_windows.zip /home/"$username"/
-sudo chown "$username":"$username" /home/"$username"/files_for_windows.zip
+if [ -f "/opt/lme/files_for_windows.zip" ]; then
+    sudo cp /opt/lme/files_for_windows.zip /home/"$username"/
+    sudo chown "$username":"$username" /home/"$username"/files_for_windows.zip
+else
+    echo "Warn: files_for_windows.zip does not exist. Probably because the LME install requires a reboot."
+fi
